move fileInputTheme to separate file
